+++ src/calendarjs.js
@@ -1,5 +1,5 @@
 /*
- * Calendar.js Library v1.3.1
+ * Calendar.js Library v1.3.2
  *
  * Copyright 2021 Bunoon
  * Released under the GNU AGPLv3 license
@@ -7041,7 +7041,7 @@ function calendarJs( id, options, startDateTime ) {
      * @returns     {string}                                                The version number.
      */
     this.getVersion = function() {
-        return "1.3.1";
+        return "1.3.2";
     };