+++ CHANGES.md
@@ -3,6 +3,7 @@
 ### Version 1.3.1:
 - Fixed a fault that caused the repeating event images to have an :hover and :active behavior.
 - Added "Hide Day" context menu support for the day headers (when more than 1 day is visible).
+- The "useOnlyDotEventsForMainDisplay" setting now overrides the "maximumEventsPerDayDisplay" setting.
 - 
 
 ### Version 1.3.0: